refactor(checker): trim type alias variable alias module
